Fix the setup.sh and cleanup.sh scripts, for easier testing

+++ cleanup.sh
@@ -1,10 +1,13 @@
-./xenvm.native shutdown
+set -ex
+
+if [ "$EUID" -ne "0" ]; then
+  echo "Please run me as uid 0. I need to create a loop device and device mapper devices"
+  exit 1
+fi
+
+./xenvm.native lvchange -an /dev/djstest/live || true
+#./xenvm.native shutdown /dev/djstest
 killall xenvmd.native
-LVS="live"
-for i in ${LVS}; do
-  echo Removing $i
-  rm -f ./djstest-$i
-done
 dmsetup remove_all
 dd if=/dev/zero of=/dev/loop0 bs=1M count=128
 losetup -d /dev/loop0
